+++ src/providers/NetworkTreeProvider.ts
@@ -3,6 +3,12 @@ import { HlfTreeItem } from '../views/trees/HlfTreeItem';
 import { NetworkItem } from '../views/trees/NetworkItem';
 import { OrganizationItem } from '../views/trees/OrganizationItem';
 import { HlfProvider } from './HlfProvider';
+import { Settings } from '../utilities/Constants';
+import { OrganizationRoot } from '../views/trees/OrganizationRoot';
+import { CaItem } from '../views/trees/CaItem';
+import { OrdererItem } from '../views/trees/OrdererItem';
+import { PeerItem } from '../views/trees/PeerItem';
+import { NodeItem } from '../views/trees/NodeItem';
 
 export class NetworkTreeProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
 
@@ -23,10 +29,21 @@ export class NetworkTreeProvider implements vscode.TreeDataProvider<vscode.TreeI
         //Hardcode the Fabric organization structure for now. Need to make this dynamic when we enable support for multiple organizations
         if (element) {
             if (element instanceof NetworkItem) {
-                tree.push(new OrganizationItem("Organizations", vscode.TreeItemCollapsibleState.Expanded));
-            }
-            if (element instanceof OrganizationItem) {
-                tree.push(new vscode.TreeItem("Org1", vscode.TreeItemCollapsibleState.None));
+                tree.push(new OrganizationRoot("Organizations", vscode.TreeItemCollapsibleState.Expanded));
+            } else if (element instanceof OrganizationRoot) {
+                tree.push(new OrganizationItem(Settings.singleOrgSettings.name, vscode.TreeItemCollapsibleState.Expanded));
+            } else if (element instanceof OrganizationItem) {
+                tree.push(new NodeItem("ca", vscode.TreeItemCollapsibleState.Expanded));
+                tree.push(new NodeItem("orderer", vscode.TreeItemCollapsibleState.Expanded));
+                tree.push(new NodeItem("peer", vscode.TreeItemCollapsibleState.Expanded));
+            } else if (element instanceof NodeItem && element.label === "ca") {
+                tree.push(new CaItem(Settings.singleOrgSettings.caDomain, vscode.TreeItemCollapsibleState.None));
+            } else if (element instanceof NodeItem && element.label === "orderer") {
+                tree.push(new OrdererItem(Settings.singleOrgSettings.ordererDomain, vscode.TreeItemCollapsibleState.None));
+            } else if (element instanceof NodeItem && element.label === "peer") {
+                tree.push(new PeerItem(Settings.singleOrgSettings.peerDomain, vscode.TreeItemCollapsibleState.Expanded));
+            } else if (element instanceof PeerItem) {
+                tree.push(new vscode.TreeItem(Settings.defaultChaincodeId, vscode.TreeItemCollapsibleState.None));
             }
         }
         else{