+++ src/utilities/Constants.ts
@@ -16,15 +16,6 @@ export enum LogType {
 
 export class Settings {
   private static _defaultChaincodeId: string = "asset";
-  static readonly debugEnv = {
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    "CORE_CHAINCODE_ID_NAME" : "asset:v1", "CORE_CHAINCODE_LOGLEVEL" : "debug", "CORE_PEER_TLS_ENABLED" : "false"
-  };
-
-  static readonly debugCaasEnv = {
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    "CHAINCODE_ID" : "asset:v1", "CHAINCODE_SERVER_ADDRESS" : "localhost:5999"
-  };
 
   //Currently peer address is hardcoded to the below and will always be launched here.
   static readonly peerAddress: string = "localhost:5052";
@@ -33,20 +24,40 @@ export class Settings {
   static readonly defaultChaincodeVersion: string = "v1";
   static dockerDir: string = "";
   static isCaas: boolean = false;
-  static singleOrgProj: string = "fabric-singleorg";
+  static readonly singleOrgProj: string = "fabric-singleorg";
+  static readonly defaultChannel: string = "default";
+
+  static singleOrgSettings = {
+    caDomain: "ca.org1.debugger.com",
+    domain: "org1.debugger.com",
+    adminUser: "Org1Admin",
+    name: "Org1",
+    msp: "Org1MSP",
+    ccpFileName: "org1.json",
+    peerDomain: "peer0.org1.debugger.com",
+    ordererDomain: "orderer.debugger.com",
+    couchDbUrl: "http://localhost:5984/_utils"
+  };
 
   static set defaultChaincodeId(chaincodeName: string){
     Settings._defaultChaincodeId = chaincodeName;
   }
 
   static get defaultChaincodeId(){
-    if(Settings.isCaas){
-      return `${Settings._defaultChaincodeId}-caas`;
-    }
-    else{
       return Settings._defaultChaincodeId;
-    }
   }
+
+  static readonly debugEnv = {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    "CORE_CHAINCODE_ID_NAME" : "asset:v1", "CORE_CHAINCODE_LOGLEVEL" : "debug", "CORE_PEER_TLS_ENABLED" : "false", "CORE_PEER_LOCALMSPID" : Settings.singleOrgSettings.msp
+  };
+
+  static readonly debugCaasEnv = {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    "CHAINCODE_ID" : "asset:v1", "CHAINCODE_SERVER_ADDRESS" : "localhost:5999", "CORE_PEER_LOCALMSPID" : Settings.singleOrgSettings.msp
+  };
+
+  static spdrLinkMessageShown: number = 0;
 }
 
 export enum DockerComposeFiles {
@@ -58,7 +69,8 @@ export enum Links {
   documentation = 'https://github.com/spydra-tech/fabric-debugger/blob/main/README.md',
   contribute = 'https://github.com/spydra-tech/fabric-debugger',
   reportIssue = 'https://github.com/spydra-tech/fabric-debugger/issues',
-  contactUs = 'https://www.spydra.app/contact-us?utm_source=vs_marketplace&utm_medium=fabric_debugger_plugin',
+  contactUs = 'https://www.spydra.app/schedule-a-demo?utm_source=vs_marketplace&utm_medium=fabric_debugger_plugin',
   review = 'https://marketplace.visualstudio.com/items?itemName=Spydra.hyperledger-fabric-debugger&ssr=false#review-details',
-  spydra = 'https://spydra.app'
+  spydra = 'https://www.spydra.app?utm_source=vs_marketplace&utm_medium=fabric_debugger_plugin',
+  getStarted = 'https://console.spydra.app/?utm_source=vs_marketplace&utm_medium=fabric_debugger_plugin'
 }