+++ src/views/HlfResponseWebview.ts
@@ -27,7 +27,7 @@ export class HlfResponseWebview {
             'fabricResponse',
             'Fabric Response',
             vscode.ViewColumn.Two,
-            {}
+            {enableFindWidget: true, enableScripts: true}
           );
     }
 
@@ -40,22 +40,30 @@ export class HlfResponseWebview {
         if(message.startsWith("Error:")){
             const statusIndex = message.indexOf("response:");
             const resultIndex = message.indexOf("message:");
-            this._message = message.substring(0, statusIndex-1);
-            this._status = message.substring(statusIndex+9, resultIndex-1);
+            if(statusIndex !== -1){
+                this._message = message.substring(0, statusIndex-1);
+            } else {
+                this._message = "";
+            }
+            if(statusIndex !== -1 && resultIndex !== -1){
+                this._status = message.substring(statusIndex+9, resultIndex-1);
+            } else {
+                this._status = " status:500";
+            }
             this._result = message.substring(resultIndex);
             this._headlineColor = "Tomato";
             if(message.indexOf("Usage:")>-1){
                 this._result = this._result.substring(0, message.indexOf("Usage:")-1);
             }
+            this._result = `<code>${this._result}</code>`;
         }
         else{
-            this._message = "Chaincode invoke successful.";
+            this._message = "Chaincode request successful.";
             this._status = " status:200";
             this._headlineColor = "MediumSeaGreen";
             if(this.isJson(message)){
                 const json = JSON.parse(message);
-                const prettyHtml = require('json-pretty-html').default;
-                this._result = `${prettyHtml(json, json.dimensions)}`;
+                this._result = `<pre><code>${JSON.stringify(json, null, 2)}</code></pre>`;
             }
             else{
                 this._result = `<code>${message}</code>`;
@@ -82,51 +90,18 @@ export class HlfResponseWebview {
                 <meta charset="UTF-8">
                 <meta name="viewport" content="width=device-width, initial-scale=1.0">
                 <title>Response</title>
-                <style>
-                    body {
-                        font-family: Menlo, Monaco, "Courier New", monospace;
-                        font-weight: normal;
-                        font-size: 14px;
-                        line-height: 16px;
-                        letter-spacing: 0;
-                        text-align: left;
-                        border-top: 1px solid #121516;
-                        padding-top: 10px;
-                        padding-bottom: 10px;
-                        margin: 0;
-                    }
-                    .json-pretty {
-                        padding-left: 30px;
-                        padding-right: 30px;
-                    }
-                    .json-selected {
-                        background-color: rgba(139, 191, 228, 0.19999999999999996);
-                    }
-                    
-                    .json-string {
-                        color: #ff6600;
-                    }
-                    
-                    .json-key {
-                        color: #3399ff;
-                    }
-                    
-                    .json-boolean {
-                        color: #00bfff;
-                    }
-                    
-                    .json-number {
-                        color: #00ff80;
-                    }
-                </style>
+                <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/default.min.css">
+                <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
             </head>
             <body>
                 <p>
                     <span style="color:${this._headlineColor};"><strong>${this._message}</strong></span><br>
                     <span><strong>Response:</strong>${this._status}</span><br>
-                    <span><strong>Time:</strong>${duration} ms</span>
+                    <span><strong>Time:&nbsp;</strong>${duration} ms</span>
                 </p>
                 <span><strong>Result:</strong><br>${this._result}</span>
+
+                <script>hljs.highlightAll();</script>
             </body>
             </html>`;
     }