+++ src/views/trees/CaItem.ts
@@ -1,0 +1,6 @@
+import { ThemeIcon } from 'vscode';
+import { HlfTreeItem } from './HlfTreeItem';
+
+export class CaItem extends HlfTreeItem {
+    contextValue = 'ca';
+}