+++ src/views/trees/OrdererItem.ts
@@ -1,0 +1,6 @@
+import { ThemeIcon } from 'vscode';
+import { HlfTreeItem } from './HlfTreeItem';
+
+export class OrdererItem extends HlfTreeItem {
+    contextValue = 'orderer';
+}