+++ src/views/trees/PeerItem.ts
@@ -1,0 +1,6 @@
+import { ThemeIcon } from 'vscode';
+import { HlfTreeItem } from './HlfTreeItem';
+
+export class PeerItem extends HlfTreeItem {
+    contextValue = 'peer';
+}