+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.0.0] - 2023-12-04
+
+- Major release with upgrade of local network to Fabric v2.5
+- Added feature - chaincode event streaming
+- Added support to connect directly from custom applications to the chaincode
+- Exposed CouchDB for debugging purposes
+- Reliability improvements
+
 ## [0.1.3] - 2023-02-05
 
 - Added support for building Typescript chaincode automatically during debug